fix: app flickers on new notification Reading updated notification state at top of `App` was triggering a re-render
